pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["sargam-poudel <sargampoudel100@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dnspython = "1.16.0"
PyNaCl = "1.3.0"
async-timeout = "3.0.1"
"discord.py" = "^1.7.3"
jishaku = "^2.3.0"
python-dotenv = "^0.19.1"
discord-custom-help = "^0.0.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.10b0", allow-prereleases = true}
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
